requirements.txt
fastapi
uvicorn
pymongo

numpy
pandas
opencv-python
pillow
pytesseract
pdf2image
pymupdf

scikit-learn
scipy

torch
torchvision

shapely
matplotlib
requests
tqdm
